+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/component/ImplicitFile.kt
@@ -4,19 +4,33 @@ import eu.vendeli.tgbot.utils.common.safeCast
 import eu.vendeli.tgbot.utils.serde.ToStringSerializer
 import kotlinx.serialization.Serializable
 
-@Serializable(ImplicitFile.Companion::class)
+@Serializable(ImplicitFile.Serde::class)
 sealed class ImplicitFile {
     abstract val file: Any
 
     @Serializable
-    class Str(
+    class FileId(
         override val file: String,
     ) : ImplicitFile()
 
     @Serializable
-    class InpFile(
+    class FileData(
         override val file: InputFile,
     ) : ImplicitFile()
 
-    internal companion object : ToStringSerializer<ImplicitFile>({ this.safeCast<Str>()?.file.toString() })
+    companion object {
+        @Suppress("FunctionName")
+        @Deprecated("Use FileData instead", replaceWith = ReplaceWith("FileData(file)"))
+        fun Inp(
+            file: InputFile,
+        ) = FileData(file)
+
+        @Suppress("FunctionName")
+        @Deprecated("Use FileId instead", replaceWith = ReplaceWith("FileId(file)"))
+        fun Str(
+            file: String,
+        ) = FileId(file)
+    }
+
+    internal object Serde : ToStringSerializer<ImplicitFile>({ this.safeCast<FileId>()?.file.toString() })
 }